rearranged post and get requests into routing folders/files
