+++ Develop/controllers/api/imageRoutes.js
@@ -1,0 +1,19 @@
+const router = require('express').Router();
+const { getImageUrlsFromFolder, getRandomImage } = require('../../utils/utils'); 
+
+// get random image URLs for each button type
+router.get('/getImageUrls', (req, res) => {
+  const healthyImages = getImageUrlsFromFolder('HealthyFood');
+  const emptyImages = getImageUrlsFromFolder('EmptyFood');
+  const badImages = getImageUrlsFromFolder('BadFood');
+  const revealImages = getImageUrlsFromFolder('RevealFood');
+
+  res.json({
+    healthy: getRandomImage(healthyImages),
+    empty: getRandomImage(emptyImages),
+    bad: getRandomImage(badImages),
+    reveal: getRandomImage(revealImages)
+  });
+});
+
+module.exports = router;
