+++ Develop/controllers/api/userRoutes.js
@@ -2,6 +2,9 @@ const router = require('express').Router();
 const { User } = require('../../models');
 const bcrypt = require('bcrypt');
 
+
+
+
 router.post('/', async (req, res) => {
     try {
       const userData = await User.create(req.body);
