+++ Develop/data.json
@@ -1,1 +1,1 @@
-{"foodLevel":11,"mood":0.14723414182662964,"size":0}
+{"foodLevel":15,"mood":0.7743722796440125,"size":4}
