+++ Develop/public/css/parentalstyle.css
@@ -12,6 +12,7 @@
   height: 0;
 }
 
+
 .slider {
   position: absolute;
   cursor: pointer;
