+++ Develop/public/css/style.css
@@ -1,6 +1,3 @@
-* {
-    /* outline: black solid 2px; */
-}
 
 header {
     width: 600px;
@@ -138,11 +135,6 @@ main {
 /*speed and how many times Mikey "spins" */
 .spinning {
     animation: spin 1s linear 2;
-}
-
-/* display Mikey upside down */
-.upside-down {
-    /* transform: rotate(90deg);  */ /* removed for now*/
 }
 
 /* keyframes for SVG spin animation  */
