+++ Develop/views/game.handlebars
@@ -1,12 +1,12 @@
 <body>
     <div id="game-menu-button-container">
     <a id="return-home" href="/">
-    <img class="custom-button" src="../images/Buttons/logout.png">
-    <img class="custom-button" src="../images/Buttons/return home.png">
+    <img class="custom-button" src="/images/Buttons/logout.png">
+    <img class="custom-button" src="/images/Buttons/return home.png">
     </a>
     </div>
     <div id="mikey-border">
-    <img id="mikey" src="../images/Mikey Frames/Mikey Frame 1.svg" alt="Mikey" />
+    <img id="mikey" src="/images/Mikey Frames/Mikey Frame 1.svg" alt="Mikey" />
     </div>
     <div id="buttonContainer">
         <button id="feedButtonBad" class="button" data-title="Feed Bad">
