+++ Develop/views/homepage.handlebars
@@ -1,27 +1,27 @@
 <body>
 <div id="homepage-main">
     <h1 id="randomStatement"></h1>
-    <img id="welcome-mikey" src="../images/Mikey Gifs/homepage.gif" alt="cartoon mouse" />
+    <img id="welcome-mikey" src="/images/Mikey Gifs/homepage.gif" alt="cartoon mouse" />
 </div>
 <div id="home-button-container">
     <!-- Conditionally render login or logout links -->
     {{#if logged_in}}
     <a href="/logout">
-    <img src="../images/Buttons/logout.png">
+    <img src="/images/Buttons/logout.png">
     </a>
     {{else}}
     <a href="/login">
-    <img id="login-button" class="custom-button" src="../images/Buttons/login-home.png">
+    <img id="login-button" class="custom-button" src="/images/Buttons/login-home.png">
     </a>
     {{/if}}
     <a href="/game">
-    <img id="play-button" class="custom-button" src="../images/Buttons/play button.png">
+    <img id="play-button" class="custom-button" src="/images/Buttons/play button.png">
     </a>
     
 </div>
     <script src="https://code.jquery.com/jquery-3.7.1.min.js"></script>
 
-    <script src="../js/script.js"></script>
+    <script src="/js/script.js"></script>
 
 
 </body>
