+++ Develop/views/layouts/main.handlebars
@@ -7,7 +7,7 @@
   <title>Hungry Mikey</title>
   <link rel="icon" type="image/x-icon" href="../../images/favicon/favicon.ico">
   <link rel="stylesheet" href="/css/jass.css">
-  <link rel="stylesheet" href="../css/style.css">
+  <link rel="stylesheet" href="/css/style.css">
   <link rel="preconnect" href="https://fonts.googleapis.com">
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
   <link href="https://fonts.googleapis.com/css2?family=Single+Day&family=Sixtyfour&display=swap" rel="stylesheet">
